Change methods to helper methods
